added health and manabar to pieces, piece can attack

+++ Assets/Scripts/Pieces/PieceDetail.cs
@@ -9,8 +9,8 @@ public class PieceDetail : ScriptableObject {
     public Sprite portrait;
     public Mesh pieceModel;
 
-    public int health;
-    public int damage;
+    public float health;
+    public float damage;
     public int armor;
 
     public int price;
